[BE/fix] version counter 에러 해결

+++ backend/src/main/java/com/jackpot/narratix/domain/repository/QnARepository.java
@@ -42,5 +42,5 @@ public interface QnARepository {
 
     List<QnA> findByIds(List<Long> qnAIds);
 
-    int incrementVersion(Long qnAId, int size);
+    long incrementVersion(Long qnAId, int size);
 }
